Sketched out models to support 'liking' a question/concern (Comment).
I added a responses field to the Comment model that maps a comment to
users who responded.  I created a CommentResponses class which implements
the through paramenter of the ManyToMany relationship in order to specify
how the user responded to the comment.

All of the naming and language is subject to change.  I just wanted to
sketch out what this might look like.