Broke custom validation error messages out into module level constants.
I wanted to be able to catch these specific types of errors in my API views,
but didn't want to rewrite the logic shane had already done for the normal
form handling.  Making the strings that will populate form.errors accessible
to the calling code seemed like the easiest way to do this.