Added a message bar and javascript function to display messages in this bar.
Javascript function is displayMessage('Message text', 'error') where the second parameter should be 'error', 'warning', 'info' or something like that.  It gets appended to the #messageswrap as a class in case we want to style different kinds of messages differently.

I also updated the login form submission to display an error message in the bar on login failure.