Added login check to handleReplyform() javascript.
This will cause an error message to be displayed when someone clicks the 'Answer this', or 'Reply' links.

In order to implmenent this, I had to add functions to check whether the user was logged in.  I decided to do this in the base template by using jQuery.data() to set a flag (is_authenticated) that can be read by other javascript code.  I wrapped the check in userIsAuthenticated() so we can modify the way that we do this check in the future without changing code.

TODO: The form is still displayed because of the handlers for .replyformtoggle.  Perhaps these handlers need to be made into functions rather than being bound to the click event.  These functions could then be called from within handleReplyform().