Added a num_responses method to the Comment model.
This method gets the number of responses of a certain type (passed as a parameter) to a given comment.  I wrote it to be used to determine the average number of responses for a topic's questions on the way to figuring out which questions are burning questions.

The query I used is to get this count is the same as num_upvotes, it can just take a parameter.  I modified num_upvotes to be a call (with response_type="concur") to num_responses.