Cleaned up attribution for clips on topic pages.
Broke template markup for displaying clips on a topic page into its own template (core-topic-answerclip.html) because it was replicated in two places in the core-topic-questions.html.

I removed the word "Citation" and put the attribution before the clip.  I also added logic to hide the date if its not provided.