Made core.Topic methods returning information on which comments users have responded to return lists of ids rather than objects.
This change in functionality comes because I initially misunderstood how the values_list() method of a QuerySet works.  From the Django docs:

"If you have a field called foo that is a ForeignKey, the default values() call will return a dictionary key called foo_id, since this is the name of the hidden model attribute that stores the actual value (the foo attribute refers to the related model)."

It caused the methods to be renamed as well.