Added logger calls when handling exceptions
I added logger calls to all the except blocks throughout the code.  I made them logger.exception() when it seemed like the code was catching an unexpected exception and logger.warning() when the code handled the code gracefully (and often returned an error to the user) but where we might want to be aware of the situation in the early stages (e.g. login failures).